requirement.txt
vpython==7.6.4
pyserial==3.5
